add scale test validations
