+++ operator/e2e/tests/scale/scale_up_test.go
@@ -127,6 +127,9 @@ func runScaleUpTest(t *testing.T, v scaleUpVariant) {
 		timeout:      scaleUpTimeout,
 		pollInterval: defaultScalePollInterval,
 	}, func(tracker *measurement.TimelineTracker, tc *testctx.TestContext, _ string) {
+		baseline := &operatorBaseline{}
+		addOperatorBaselinePhase(tracker, tc, baseline)
+
 		var deployMilestones []measurement.MilestoneDefinition
 		if v.initialPods > 0 {
 			deployMilestones = append(deployMilestones,
@@ -189,21 +192,10 @@ func runScaleUpTest(t *testing.T, v scaleUpVariant) {
 			},
 		})
 
-		tracker.AddPhase(measurement.PhaseDefinition{
-			Name: "delete",
-			ActionFn: func(ctx context.Context) error {
-				return workload.NewWorkloadManager(tc.Client, Logger).DeletePCS(ctx, tc.Namespace, tc.Workload.Name)
-			},
-			Milestones: []measurement.MilestoneDefinition{
-				{
-					Name: "pcs-deleted",
-					Condition: &condition.PCSDeletedCondition{
-						Client:    tc.Client.Client,
-						Name:      tc.Workload.Name,
-						Namespace: tc.Namespace,
-					},
-				},
-			},
-		})
+		addFinalCheckAndDeletePhases(tracker, tc, scaleFinalCheckConfig{
+			targetReplicas: v.targetReplicas,
+			targetPods:     v.targetPods,
+			workloadName:   v.workloadName,
+		}, baseline)
 	})
 }
